Call on render thread working via CommandBuffer, D3D11::Map working, Unmap doesn't seem to apply
Translate C# job works, the render thread calls our native function with the data ptr, which then uploads and allows us to translate again.
Map works now the mesh is marked as dynamic,
However, changes do not seem to be applied, are we writing into a copy of the mesh? No changes visible in game view.