InputState copied to C++ State in PreExecute and then consumed, InputManager
New InputState struct that stores all input related data that libigl needs, flags, input values etc.
We copy this into the State.Input in PreExecute and then immediately consume it with ConsumeInput so we can detect new changes.
Makes input accessible from C++ as well
InputManager handles controller models as well as getting the controllers.
Getting the actual input state is done in the partial class Behaviour.Input.cs
Anything that may change as we are executing should be in the InputState